requirements.txt
# =============================================================================
# JurisAI - Phase 1: Fine-Tuning Dependencies
# =============================================================================

# Core ML Framework
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0

# Hugging Face Ecosystem
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.13.0
bitsandbytes>=0.44.0
datasets>=3.0.0
huggingface-hub>=0.25.0
tokenizers>=0.20.0
sentencepiece>=0.2.0
trl>=0.11.0

# Unsloth - Fast QLoRA Training
unsloth

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0.0
jsonlines>=4.0.0

# Text Processing
regex>=2024.0.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.9.0
scikit-learn>=1.5.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.7.0

# Logging
tensorboard>=2.17.0
